r31_20090312_0433_13751_b40895fb6101dbe96dc7b9d6477eeeb4

+++ hh_room_utils.cct/Dialog Thread Class.ls
@@ -247,10 +247,15 @@ on createDialog me, tWndTitle, tWndType, tContentType, tEventProc
 end
 
 on removeDialog me, tWndTitle, tWndList
+  if not listp(tWndList) then
+    return 0
+  end if
   if tWndList.getOne(tWndTitle) then
     tWndList.deleteOne(tWndTitle)
-    if not voidp(pUrlList.getaProp(tWndTitle)) then
-      pUrlList.deleteProp(tWndTitle)
+    if pUrlList.ilk = #propList then
+      if not voidp(pUrlList.getaProp(tWndTitle)) then
+        pUrlList.deleteProp(tWndTitle)
+      end if
     end if
     return removeWindow(tWndTitle)
   else
